Fix resource group conditional to 5.2.0
Previously, the conditional was supposed to be set to 5.20.0 because the
propose variables were to be removed. Since then we have decided it was
a breaking change and would be left in. In this commit we change the
resource group conditional to match when pg_dumpall was modified to use
value instead of propose variables in GPDB commit 65df77c5edc167ec.

Co-authored-by: Kevin Yeap <kyeap@pivotal.io>
Co-authored-by: Jimmy Yih <jyih@pivotal.io>