Fix gprestore verbose output when using --jobs flag
When the --jobs flag is given, gprestore would output "(table 1 of N)"
multiple times depending on the --jobs input. The last table to be
restored would also not output the expected "(table N of N)". This is
because the first restore for each parallel worker gets the same table
number (in this case 1) which will be outputted before restoring
data. After the restore is finished, each parallel worker atomically
increments the table number.

To fix the incorrect gprestore output, we initialize the table number
to 0 and have each parallel worker atomically increment the table
number after they are done with their restore.