Cancel blocked LOCK TABLE query during gpbackup DoCleanup
The gpbackup utility can leak a blocked SQL session when a user sends
SIGINT (Ctrl-C) while gpbackup is executing its LOCK TABLE query. This
only occurs when the query is blocked trying to obtain a simple
AccessShareLock on a table when another session has already acquired
an AccessExclusiveLock on that same table. The AccessExclusiveLock
could have been acquired by the other session explicitly via LOCK
TABLE or implicitly via commands like TRUNCATE and ALTER TABLE.

The reason the SQL session is leaked is because the Golang SQL
libraries do not cancel the queries for you when closing
connections. To do this, we must introduce the concept of Golang
Contexts which gives us a very basic cancel function we can call
during clean up. The query cancel itself does not block and there is
no cancel verification built in so we explicitly queue a ROLLBACK on
the main gpbackup SQL session to make sure the SIGINT makes it through
the libpq connection. The Context and its cancel function are stored
as globals for now as one-off variables but I can see it grow into a
big list since all other gpbackup queries could technically encounter
this too (but unlikely).