Adds test for NOT VALID clause
Gpbackup supports NOT VALID option on constraints. There is an edge case
where a table can contain data that violates it's constraints. This can
happen when data is inserted into a table before putting a constraint on
the table.  This commit adds a behavior test to make sure we can restore
constrained tables with legacy constraint-violating data.

Reference Commit:
https://github.com/greenplum-db/gpbackup/commit/90b3b54b83fe01f8107a329f2bd10652abd67bd6

Authored-by: Ekaterina Dontsova <edontsova@pivotal.io>