Run make depend in 5X-head-gpbackup-fixed-test CI job
The job used to use the CI-cached dependencies of the latest gpbackup
which was fine for a while since it was only compiling tests. However,
a recent change updated the DATA-DOG/go-sqlmock test dependency so now
we can no longer use the CI-cached dependencies for these old
tests. As a quick fix, we just need to run make depend before running
any tests. One possible improvement could be to cache the old
dependencies so we do not need to always install dep and pull
dependencies from Github (or at least install dep on CCP images).

Reference Commit:
https://github.com/greenplum-db/gpbackup/commit/c1caad1634297ec0

Co-authored-by: Kevin Yeap <kyeap@pivotal.io>