Support materialized views
Add backup support for materialized views in GPDB6.1. Since they are
closely related to views, the query was reused and GetViews was changed
to also query information about materialized views. The reason they are
then sorted into separate object types is to make it easier to implement
materialized views ACL.

Co-authored-by: Kevin Yeap <kyeap@pivotal.io>
Co-authored-by: Shivram Mani <shivram.mani@gmail.com>