Remove manual dependency caching
Dependency caching was originally implemented as a workaround to reduce
flakes caused by github.com rate limiting. With the switch to go modules
and the upgrade to Go 1.13.4, we can now take advantage of the GOPROXY
flag. The GOPROXY flag as of Go 1.13 defaults to
"https://proxy.golang.org,direct". This means dependency download will
try looking for dependencies in Google's proxy before going to github.
If we feel the need to implement manual dependency caching again. We can
vendor using go modules and build with the "-mod=vendor" flag.

Authored-by: Kevin Yeap <kyeap@pivotal.io>