Handle backup of child partitions with different schema than the root
It is possible to have child partitions existing in different schemas
than their root partition's schema. Without logic to handle this
scenario, gpbackup was creating metadata that would put those child
partitions back onto the same root partition schema. Fix this issue by
keeping track of which child partitions have had their schema modified
and dump ALTER TABLE SET SCHEMA statements for them.

One big thing to note is that this is a bit complex when involving the
--leaf-partition-data flag and/or the schema/table filtering flags. If
a filter is given which would exclude the schema of the child
partition, we will make an exception to the filter and add the CREATE
SCHEMA statement(s) needed to guarantee table/catalog integrity for
the partition table DDL.