Fix some psql calls when using .psqlrc file
After the Postgres 9.6 merge into GPDB master, the .psqlrc file is
always parsed if the -X flag is not given to psql. This will make any
psql call to also print out the .psqlrc file parsing. We have a couple
areas where we depend on only obtaining the query output but this
extra parsing text is breaking the logic. Adding the -X flag will make
it always work as expected.

GPDB commit reference:
https://github.com/greenplum-db/gpdb/commit/5a92cefbcf35b8bc