Fix --redirect-schema when restoring to original database
The previous --redirect-schema tests did not account for the main use
case of redirecting back into the original database that is expected to
still contain the original tables.  When --redirect-db is not specified,
gprestore will validate the tables we want to restore into do not exist.
When --redirect-schema is being used without --redirect-db, check
redirected tables do not exist.

gpbackup reference commit:
https://github.com/greenplum-db/gpbackup/commit/35e8262e59adb02e

Authored-by: Kevin Yeap <kyeap@pivotal.io>