Fix incremental backup end_to_end test
The test started failing after a formatting commit that incorrectly
refactored a portion of the test which checks for two distinct aoseg
entries for the test table. In GPDB 4, 5, and 6, the gp_toolkit
function returns the entries from the master segment.  In GPDB 7+, the
gp_toolkit function returns all entries from the primary segments.

gpbackup commit reference:
https://github.com/greenplum-db/gpbackup/commit/60b6c00baf8dbfa3