Create savepoint before getting view definition to avoid locking (#375)
When querying the view definition using pg_get_viewdef(), this function
obtains dependency locks that are not released until the transaction is
committed at the end of gpbackup session. This blocks other sessions
from commands that need AccessExclusiveLock (e.g. TRUNCATE)

Creating a SAVEPOINT allows the locks to be released once the savepoint
is released or rolled back.