Fix pg_class statistics dump
The relnamespace part of the UPDATE query cannot be the schema's
actual oid. The UPDATE query would always report 0 rows updated for
the majority of cases and only work if restoring to a database with
that exact schema oid for that exact schema name. This was not caught
in production because gp_autostats_mode GUC default setting would
create the statistics for pg_class at the end of COPY FROM
SEGMENT. With a recent commit turning off autostats during gprestore
--with-stats, the pg_class statistics no longer get filled
accidentally. This commit fixes the issue by filtering on the pg_class
oid field and using regclass on the table's fully-qualifed name for
the UPDATE query.

Reference Commit:
https://github.com/greenplum-db/gpbackup/commit/0bf5340e4fc8b70c02361e5024819828e5eb6323