Incremental restore should filter out existing external tables
Incremental restore was not properly filtering out existing external
tables. This was because the query for finding existing tables did not
include external tables in the relkind and relstorage query
filters. Fixed by adding them to the filters. Note that 'f' for
foreign tables only matters for GPDB 6+ but we shouldn't need a GPDB
version check on this since this is catalog and 'f' is new starting
from GPDB 6+.

Commit Reference:
https://github.com/greenplum-db/gpbackup/commit/c8ab236ce88b1c94