Fix end_to_end tests
A recently added end_to_end test left an external table in the public
schema of the test database. This resulted in wrong results (# of
tables expected) for other subsequent tests. Fix by moving the
external table into the schema specific for the test the external
table is created in.

Reference Commit:
https://github.com/greenplum-db/gpbackup/commit/73685651b13c2209f303973fe35ed51b065167ec