Remove undocumented external table option attribute
After a changes to external tables in GPDB, it was discovered that the
options attribute on external tables was an undocumented feature. Since
it is undocumented it should be safe to remove. This commit also fixes
some formatting changes to format options attribute that was the result
of commits that changed how external table format options were stored.

GPDB reference commits:
https://github.com/greenplum-db/gpdb/commit/e4b499aa4a16df49241abc9dd70acec5589f1323
https://github.com/greenplum-db/gpdb/commit/eddd3a6e08e81c19b4588a3718cf1cb7197cfaf9

Authored-by: Kevin Yeap <kyeap@pivotal.io>