gprestore ability to discard or seek or scan subset data (#390)
Enhanced gprestore to optimize the way we scan the backup data during restore specifically in the scenario when we have an uncompressed single file backup with table/schema filters

Introduced RestoreReader which encapsulates both reader types io.ReadSeeker and bufio.Reader. We will use the appropriate reader based on the ReaderType. The reader can be one one of the following types
- reader that can directly seek during restore (restore from a filesystem)
- reader that can read subset data based on the target table offsets from toc (when we use plugin)
- reader that read entire data and based on offsets from toc, data will need to be discarded.

The functions to move to a specific reader offset or copy data from the reader are now functions of the RestoreReader

- Introduced a new function to the plugin interface named restore_data_subset which is optimized for reading subset data. In order to make use of this the plugin config file needs to set restore_subset: "on"
This takes in an offset file containing all the start and end byte offsets of the relevant oids that need to be restored.

- ddboost plugin tests have been added to plugin_test_bench.sh

- ddboost scale tests have been newly introduced in plugin_test_scale.sh

- s3_plugin doesn't support the new api