Fix backup_helper integration test intermittent failures
There are two tests that sleep for 200 milliseconds and then
immediately sends a SIGINT to the backup_helper. Afterwards, the test
checks for a pipe error file. This test intermittently fails when
backup_helper is too slow and hasn't gotten to the point where pipe
error files would be generated.

To fix this, a waiting function has been added to check for two pipe
files to have been created with a timeout of 5 seconds. It's peculiar
why both pipe files need to be created instead of just one but that
can be looked into at a later date.