Skip --run-analyze and --leaf-partition-data end_to_end test on GPDB 4.3
In GPDB 4.3, leaf partition statistics are not automatically merged up
to the root partition when all leaf partitions have been analyzed. The
gprestore --run-analyze feature needs to be updated to always run
ANALYZE on the root partition in GPDB 4.3. Skip the test for now when
running against GPDB 4.3 and add a TODO for tracking.