Fixes to plugin_test_bench.sh
With the previous commit d9b7d48. The timestamp is no longer on the 4th
line of gpbackup's output which is breaking our grep to get timestamp in
plugin_test_bench.sh. Increased the limit to prevent additional breakage
if we decide to add more logging. Also fixed log output.

Authored-by: Kevin Yeap <kyeap@vmware.com>