Fix plugin perf tests
The plugin perf test script extracts the backup timestamp from the log
output. The verbose log output was recently changed so the test was
not able to get the backup timestamp and fails. As per a similar fix,
just need to observe more lines of the log output to extract the
backup timestamp.

Relevant gpbackup commits:
https://github.com/greenplum-db/gpbackup/commit/8c83df0948843310959f92424da3e120038f41a4
https://github.com/greenplum-db/gpbackup/commit/d9b7d48a258687aded588ab7f137baf8290af2d6