Get rid of unnecessary trigger points in CI pipeline
There were a lot of unnecessary trigger points found in the CI
pipeline. To highlight a couple of them:
1. A nightly trigger on the build job which would trigger the entire
pipeline afterwards. There's no changes so this would only be an
expensive attempt to flush out intermittent failures.
2. Trigger points on new GPDB binaries for plugin-related jobs. These
jobs are only for testing plugin APIs for the external storage
(e.g. S3 and Data Domain). A new GPDB binary would not affect
anything.
3. Trigger points on new GPDB binaries for scale test jobs. These jobs
are specifically for testing backup/restore scaling on different
backup sizes. Although a change in the GPDB binary could cause this,
it's very unlikely and would be caught in GPDB performance regression
testing efforts instead.

One trigger point was added though for when the master job discovers a
new GPDB master binary. Although possibly noisy due to frequent GPDB
master commits, this could be a good one to have to keep up to
date. We can always remove it later if it becomes too much to handle.