Dump persistent error logs for external tables
In GPDB 5+, a user can create an external table with the option LOG
ERRORS PERSISTENTLY which persists rejected row error logs into a flat
file so that recreating the table would use the same error log. The
PERSISTENTLY part was missing from the gpbackup metadata dump which
this patch fixes.

Note: There is a hidden OPTIONS syntax for CREATE EXTERNAL TABLE in
GPDB 5+ which stores the PERSISTENTLY flag bit. We only examine the
OPTIONS text array for error_log_persistent=true and ignore any other
options since the OPTIONS syntax is an undocumented (and thus
unsupported) feature. It seems like it was a remnant part of the
gpcloud feature which has since lost developer cycles in favor of the
more supported PXF feature.