Fix end_to_end plugin test testing all objects for GPDB 7
There were some incompatible SQL in the gpdb4_objects.sql script which
do not run properly in GPDB 7. To work around this, separate out the
bad queries into their own separate SQL files and update the queries
to be compatible to GPDB 7.

A GPDB 7 TODO is left over to fix a dependency issue for external leaf
partitions.