Fix some integration tests for GPDB 7 changes
Some gpbackup commits that went in did not run against the integration
test suite which resulted in some missing test refactors to introduce
GPDB 7 specific struct changes. Also, proparallel was not being
queried correctly for GetAggregates() because it was accidentally
placed in the GPDB 6 query and not the GPDB 7 query.