Handle COPY deadlocks during gpbackup by deferring to main worker thread
The old solution which had all worker threads acquiring
AccessShareLocks was recently reverted due to GPDB shared memory
errors showing up. To replace it, we now attempt to acquire an
AccessShareLock right before running COPY by running LOCK TABLE
NOWAIT. If the LOCK TABLE NOWAIT returns an error (due to a random
external session having queued an AccessExclusiveLock acquisition
request), we catch it and apply deadlock handling logic. The logic
consists of deferring the table to a deferred queue and terminating
the worker (its transaction is now in an aborted state and the
distributed snapshot cannot be recovered). Once all tables in the
regular queue have been operated on, the main worker thread will then
start running COPY on the tables in the deferred queue.

This logic is similar to what happens in parallel pg_dump in upstream
Postgres. However, pg_dump aborts and fails when this deadlock
scenario happens... whereas we defer the table to the main worker
thread which we know will work.