Detect and panic on views with anyarray type casts in view definition
View definitions with anyarray typecasts are not restorable. If we
detect an anyarray type cast, we error out and panic to prevent a
corrupted backup that will cause gprestore to error out later. A
helpful message is given as well to make sure the user is notified
that they must manually drop the view or recreate the view without
explicit array type casts.

Related GPDB commit reference:
https://github.com/greenplum-db/gpdb/commit/5ef30dd4c9878aba