Fix terraform issues in Concourse CI pipeline
Hashicorp (terraform) rotated their GPG key used for release signing
and verification. Because of this, our old version of the terraform
Concourse plugin is unable to get the terraform cloud provider modules
needed to create the CCP VMs for CCP jobs. To get around this, we
cache the terraform cloud provider modules and use the cached modules.

To fully fix the issue instead of using this workaround, the CCP code
must be refactored to work fully with the latest terraform Concourse
plugin.

Reference:
https://discuss.hashicorp.com/t/hcsec-2021-12-codecov-security-event-and-hashicorp-gpg-key-exposure/23512
https://github.com/greenplum-db/gpdb/commit/ac21882dc5a363a38508cc203c3cd18dda0397ac