Replace docker-image with registry-image in Concourse pipelines
Our gpbackup pipeline started to encounter many "failed to stream in
to volume" failures due to some job tasks trying to use a volume
before the volume was created. As a workaround, replace the
docker-image resources with registry-image in the pipeline yaml
configuration. The registry-image resource is much more light-weight
and much less prone to errors.

Reference:
https://github.com/greenplum-db/gpdb/commit/be6ff30fb2e2a74d423b29afaf863f698d3e842d