Update filepath function ParseSegPrefix to allow error handling
Previously, when an error was encountered when parsing a backup's
segment prefix, the program would panic. ParseSegPrefix now returns
an error and the caller of the function is responsible for handling any
errors encountered. This allows for greater flexibiliy in how a program
should react in the event of an error instead of panicking everytime.

Authored-by: Kevin Yeap <kyeap@vmware.com>