Quote table names for gprestore exclude relation filter
When using the exclusion filters with gprestore, we were not properly
quoting the excluded tables (e.g. for table names that have special
characters or are keywords). The tables would bypass the exclusion
filter and get unexpectedly restored. To fix this issue, simply quote
the user-given tables in the exclusion filter similar to how we handle
the inclusion filter.

This was discovered while reviewing
https://github.com/greenplum-db/gpbackup/pull/500. In that PR, github
user tm-drtina only handled the issue for gpbackup.  This commit
handles the issue for gprestore.