Allow --redirect-schema to work with included schemas
This commit adds support to run --redirect-schema with either the
--include-schema or --include-schema-file flags. This was a small
oversight when the --redirect-schema feature was introduced.

Note that usage of this feature can result in conflicts if the user is
not careful (e.g. two schemas containing duplicate table names being
redirected into the same schema can cause a conflict). It was decided
not to have any intelligent logic to prevent/warn against these
scenarios but will be documented properly. It will be at the user's
discretion to identify any duplicates before deciding to use this
feature.