Create skip file only when restoring to GPDB 6X+
On GPDB 6X+, COPY FROM SEGMENT is more desynchronized with the
interactions between the QD and the QEs. Each QE backend executes its
COPY with minimal interactions with the QD. The QD immediately
broadcast aborts if a single QE backend reports back with an
error. Afterwards, an error is immediately reported to the
user. Because of this, the restore_helper could get into situations
where its infinitely waiting for its expected COPY reader to open the
pipe but the COPY reader had already aborted before attempting to open
the pipe... hence the skip file fix we made.

On GPDB 5X and below, the QD and QEs have much more synchronization
and will have some back and forth between each other (e.g. all QEs
will attempt to acquire lock, report back to the QD, and wait for QD
to tell it that the others have also acquired the lock). The QD also
waits for all other QE segments to finish before aborting the
transaction and reporting the error to the user. Because of the higher
amount of synchronization, the skip file implementation is not needed
for 5X and below. We disable it to make sure skip files are not
unnecessarily created and leaked.