Ensure that skip files are only created for single data file restores
The skip file logic is meant to help synchronize gprestore and its
restore_helpers. Since restore_helpers are only invoked when gprestore
is restoring a single data file backup, we should restrict the skip
file logic as such. Without it, skip files would be unnecessarily
created and leaked when gprestore --on-error-continue on a standard
backup encounters an error.