Set restore_helper pipe writer to block on IO
As part of the skip file implementation, the pipe writer was
refactored to open with O_NONBLOCK. However, there was no logic added
for the pipe writer to handle EAGAIN errors due to the pipe buffer
being full. Because of that, it became common to see the
restore_helpers failing to provide data to the COPY pipe reader
(e.g. COPY would fail due to receiving partial data cut off by the
buffer).

To fix the issue, simply set the file descriptor to block again on
IO. We only needed the O_NONBLOCK on the open logic anyways so this
patch reverts the write implementation back to what it was before.