Fix S3 Concourse jobs failing to terraform provision
The S3 Concourse jobs would intermittently fail due to the terraform
plugin not being able to find the proper provider plugins
(e.g. google, aws, etc). This was because the provider plugin cache
was being retrieved in parallel to the terraform provisioning
task. This patch makes sure the provider plugin cache is retrieved
serially before the terraform provisioning task.