Log AccessExclusiveLock held on relation when worker cannot obtain AccessShareLock
In case we aren't able to grab an AccessShareLock on the table being
backed up, we terminate the worker thread and its corresponding
connection and let the main worker thread (Worker 0) handle them. This
can happen when there is another session with an AccessExclusiveLock on
the same table being handled by the worker thread.

When this occurs, log the AccessExclusiveLock being held on the table.

Authored-by: Brent Doil <bdoil@vmware.com>