Lock tables in MASTER ONLY mode.
This commit utilitizes the MASTER ONLY lock mode
feature introduced in GPDB 6.21.0 to only take
the required AccessShareLocks on the master to
prevent DDL changes.

A combination of this commit and increasing the
max_locks_per_transaction GUC on the master
should dramatically help with shared memory OOM
issues when doing backups on very large datasets.

Authored-by: Brent Doil <bdoil@vmware.com>