Rename exchanged partition table indexes
In the case that a partitioned table has indexes, exchanging a leaf
will result in both the original leaf and the new partition table having
indexes with names matching the original table name.  This will cause restore
to fail due to a name collision.  This is only an isue on GPDB6 and
earlier.

To resolve this, we retrieve indexes meeting these criteria and rename them
to the convention matching the exchanged leaf table's name before backing up.

This situation also happens with constraint indexes on GPDB7 and above,
due to changes made to align with upstream.  An equivalent solution is
applied in this case.

Note that we use two very similar functions to handle constraints and
indexes.  This is necessary due to current limitations of generics in
Go.

Co-authored-by: Andrew Repp <reppa@vmware.com>
Co-authored-by: Jamie McAtamney <jmcatamney@vmware.com>