Prevent test hang on Concourse
The resize-cluster end-to-end test currently hangs on Concourse, for
reasons that we have been unable to determine and cannot reproduce
locally.  In order to prevent blocking pipelines on that test, this
commit adds a mechanism to kill and clean up the gprestore process when
it hangs as a workaround until we can fix the actual issue.